backend/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
ua-parser==0.18.0
uritemplate==4.1.1
user-agents==2.2.0
